FEATURE: maintenance konsolen befehl vereinheitlicht
